requirements.txt
PyYAML>=6.0
pandas>=1.5
numpy>=1.23
ib-insync>=0.9
python-dateutil>=2.8
yfinance>=0.2
zoneinfo; python_version<"3.9"
tzdata; python_version<"3.9"
